refactor: Intege -> Long으로 변환

+++ src/main/java/com/nft/reservation/domain/concert/ConcertService.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface ConcertService {
     ConcertDTO createConcert(ConcertForm concertForm);
-    ConcertDTO getConcertDetail(Integer concertId);
-    List<SeatDTO> getConcertSeat(Integer concertId);
-    ConcertHallDTO getConcertHallSize(Integer concertId);
+    ConcertDTO getConcertDetail(Long concertId);
+    List<SeatDTO> getConcertSeat(Long concertId);
+    ConcertHallDTO getConcertHallSize(Long concertId);
 }
